pyproject.toml
[project]
name = "mlx-sam3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ftfy>=6.3.1",
    "huggingface-hub>=1.1.6",
    "iopath>=0.1.10",
    "mlx>=0.30.0",
    "numpy>=2.3.5",
    "pillow>=12.0.0",
    "regex>=2025.11.3",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
]
